+++ Strava API v3/src/javastrava/util/impl/gson/serializer/MapPointSerializer.java
@@ -39,7 +39,7 @@ public class MapPointSerializer implements JsonDeserializer<StravaMapPoint>, Jso
 	public StravaMapPoint deserialize(final JsonElement element, final Type type, final JsonDeserializationContext context)
 			throws JsonParseException {
 		JsonArray array = element.getAsJsonArray();
-		return new StravaMapPoint(array.get(0).getAsFloat(), array.get(1).getAsFloat());
+		return new StravaMapPoint(Float.valueOf(array.get(0).getAsFloat()), Float.valueOf(array.get(1).getAsFloat()));
 	}
 
 }
